requirements.txt
numpy>=1.21.0
colorama>=0.4.6
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.5.0
# torch>=2.0.0  # Optional: For learning phase, install manually if available
